+++ app/config/database.php
@@ -54,10 +54,10 @@ return array(
 
 		'mysql' => array(
 			'driver'    => 'mysql',
-			'host'      => getenv('MYSQL_HOST'),
-			'database'  => getenv('MYSQL_DATABASE'),
-			'username'  => getenv('MYSQL_USERNAME'),
-			'password'  => getenv('MYSQL_PASSWORD'),
+			'host'      => 'localhost',
+			'database'  => getenv('DB_NAME'),
+			'username'  => getenv('DB_USERNAME'),
+			'password'  => getenv('DB_PASSWORD'),
 			'charset'   => 'utf8',
 			'collation' => 'utf8_unicode_ci',
 			'prefix'    => 'typi_',