+++ public/js/public/master.js
@@ -7,3 +7,22 @@ $(".fancybox").fancybox({
     openEffect: 'elastic',
     closeEffect: 'elastic'
 });
+
+/**
+ * Slider
+ */
+var mySwiper = new Swiper('.swiper-container', {
+    pagination: '.swiper-pagination',
+    loop: true,
+    grabCursor: true,
+    paginationClickable: true,
+    calculateHeight: true
+});
+$('.arrow-left').on('click', function(e){
+    e.preventDefault()
+    mySwiper.swipePrev()
+});
+$('.arrow-right').on('click', function(e){
+    e.preventDefault()
+    mySwiper.swipeNext()
+});